Répartition backend + Rafraichissement automatique
- Dans le backend, répartition en fichiers différents pour les différentes routes. Cela permet de simplifier la lecture du code et de faire la part des choses.

- Rafraichissement automatique après modification et suppression d'une réservation. L'utilisateur ne devra plus appuyer sur le bouton refresh pour voir les nouvelles modifications.

- Changement pour le rafraichissement automatique après ajout d'une réservation. Quand on cliquait pas sur le bouton Fermer après ajout d'une réservation réussie, les réservations ne s'actualisaient pas.

- Changement pour quand on revient sur une page de Schedule mais d'un autre local. Maintenant, ça va bien se refresh sans l'aide d'un windows.location.refresh (ce qui ralentissait l'application) mais grâce à un render dans la page App.tsx avec une clé contenant le nom du local.
