+++ ephecReservation/backend/database.js
@@ -1,0 +1,10 @@
+import mysql from "mysql";
+
+const db= mysql.createConnection({
+  host: "localhost",
+  user: "root",
+  password: "",
+  database:"ephecreservation"
+});
+
+export default db;
